Add point to both sides of trends line graphs
This centers the actually relevant data points within sections when the
number of points and the preferred number of sections are the same. It
makes no sense from a logical or data representation point of view,
since the graph extends from edge to edge. When there are more points
than sections, the change is indistinguishable, as section labels are
selected from the center of each section instead of the first value.