Add debug option to change API address
Presents the user with an alert view to change the current API server
address. The change is reset upon each launch.

This is a feature to facilitate design, development, and QA teams being
able to view more variations of API data than their own and thus detect
design inconsistencies or implementation bugs.